Fix bug for deleting yml file

+++ app.py
@@ -87,10 +87,12 @@ def main():
                      }
 
             # Configuration file
+            PEP8SPEAK_YML_FOUND = False
             r = requests.get("https://api.github.com/repos/" + \
                             repository + "/contents/").json()
             for content in r:
                 if content["name"] == ".pep8speaks.yml":
+                    PEP8SPEAK_YML_FOUND = True
                     res = requests.get(content["download_url"])
                     with open(".pep8speaks.yml", "w+") as config_file:
                         config_file.write(res.text)
@@ -114,7 +116,8 @@ def main():
             except:  # Bad yml file
                 pass
 
-            os.remove(".pep8speaks.yml")
+            if PEP8SPEAK_YML_FOUND:
+                os.remove(".pep8speaks.yml")
 
             # Run pycodestyle
             r = requests.get(diff_url)
